+++ src/hooks/useColumns.tsx
@@ -18,38 +18,38 @@ export const useColumns = <T extends any>(
   return useMemo<Column<T, any>[]>(() => {
     const partialColumns: Partial<Column<T, any>>[] = [
       {
-        width: '0 0 40px',
-        minWidth: 0,
-        title: <div className="dsg-corner-indicator" />,
-        component: defaultGutterComponent,
-        isCellEmpty: cellAlwaysEmpty,
         ...gutterColumn,
+        width: gutterColumn?.width ?? '0 0 40px',
+        minWidth: gutterColumn?.minWidth ?? 0,
+        title: gutterColumn?.title ?? <div className="dsg-corner-indicator" />,
+        component: gutterColumn?.component ?? defaultGutterComponent,
+        isCellEmpty: cellAlwaysEmpty,
       },
       ...columns,
     ]
 
     if (stickyRightColumn) {
       partialColumns.push({
-        width: '0 0 40px',
-        minWidth: 0,
         ...stickyRightColumn,
+        width: stickyRightColumn.width ?? '0 0 40px',
+        minWidth: stickyRightColumn.minWidth ?? 0,
         isCellEmpty: cellAlwaysEmpty,
       })
     }
 
     return partialColumns.map<Column<T, any>>((column) => ({
-      width: 1,
-      minWidth: 100,
-      renderWhenScrolling: true,
-      component: defaultComponent,
-      disableKeys: false,
-      disabled: false,
-      keepFocus: false,
-      deleteValue: identityRow,
-      copyValue: defaultCopyValue,
-      pasteValue: identityRow,
-      isCellEmpty: defaultIsCellEmpty,
       ...column,
+      width: column.width ?? 1,
+      minWidth: column.minWidth ?? 100,
+      renderWhenScrolling: column.renderWhenScrolling ?? true,
+      component: column.component ?? defaultComponent,
+      disableKeys: column.disableKeys ?? false,
+      disabled: column.disabled ?? false,
+      keepFocus: column.keepFocus ?? false,
+      deleteValue: column.deleteValue ?? identityRow,
+      copyValue: column.copyValue ?? defaultCopyValue,
+      pasteValue: column.pasteValue ?? identityRow,
+      isCellEmpty: column.isCellEmpty ?? defaultIsCellEmpty,
     }))
   }, [gutterColumn, stickyRightColumn, columns])
 }